+++ src/main/java/techcourse/myblog/web/controller/UserController.java
@@ -9,14 +9,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.servlet.view.RedirectView;
-import techcourse.myblog.domain.User;
 import techcourse.myblog.dto.UserDto;
 import techcourse.myblog.service.UserReadService;
 import techcourse.myblog.service.UserWriteService;
 import techcourse.myblog.validation.UserInfo;
 import techcourse.myblog.web.support.SessionUser;
 
-import javax.servlet.http.HttpSession;
 import javax.validation.groups.Default;
 
 @Controller
@@ -46,34 +44,6 @@ public class UserController {
 
         userWriteService.save(userDto.toUser());
         return new RedirectView("/login");
-    }
-
-    @GetMapping("/login")
-    public String createLoginForm(Model model) {
-        if (!model.containsAttribute("userDto")) {
-            model.addAttribute("userDto", new UserDto("", "", ""));
-        }
-
-        return "login";
-    }
-
-    @PostMapping("/login")
-    public RedirectView login(HttpSession session,
-                              @Validated(Default.class) UserDto userDto,
-                              BindingResult bindingResult) throws LoginBindException {
-        if (bindingResult.hasErrors()) {
-            throw new LoginBindException(bindingResult);
-        }
-
-        User loginUser = userReadService.login(userDto);
-        session.setAttribute("user", loginUser);
-        return new RedirectView("/");
-    }
-
-    @GetMapping("/logout")
-    public RedirectView logout(HttpSession session) {
-        session.invalidate();
-        return new RedirectView("/");
     }
 
     @GetMapping("/users")
